moved room + user to be domani types
