+++ service/src/domain/agent/events/matrix.event.adapter.room.ts
@@ -8,7 +8,7 @@ import {
 import { CommunicationEventMessageReceived } from '@src/services/communication-adapter/dto/communication.dto.event.message.received';
 import { MatrixRoomInvitationReceived } from '@src/services/communication-adapter/dto/communication.dto.room.invitation.received';
 import { MatrixMessageAdapter } from '../../adapter-message/matrix.message.adapter';
-import { MatrixRoom } from '../../adapter-room/type/matrix.room';
+import { MatrixRoom } from '../../room/matrix.room';
 import { MatrixRoomAdapter } from '../../adapter-room/matrix.room.adapter';
 import { MatrixClient, MatrixEvent, RoomMember, KnownMembership } from 'matrix-js-sdk';
 import { MatrixEntityNotFoundException } from '@src/common/exceptions/matrix.entity.not.found.exception';
