+++ service/src/domain/matrix-admin/rooms/matrix.admin.rooms.service.ts
@@ -4,7 +4,6 @@ import { LogContext } from '@src/common/enums/logging.context';
 import { RoomPowerLevelsEventContent } from 'matrix-js-sdk/lib/types';
 import { MatrixEntityNotFoundException } from '@src/common/exceptions/matrix.entity.not.found.exception';
 import { EventType, IStateEventWithRoomId, MatrixClient } from 'matrix-js-sdk';
-import { IOperationalMatrixUser } from '../../adapter-user/matrix.user.interface';
 import { MatrixUserAdapter } from '../../adapter-user/matrix.user.adapter';
 import { MatrixAdminEventLogRoomStateInput } from './dto/matrix.admin.rooms.dto.event.log.room.state';
 import { MatrixAdminUserElevatedService } from '../user-elevated/matrix.admin.user.elevated.service';
@@ -12,6 +11,7 @@ import { MatrixAdminUserService } from '../user/matrix.admin.user.service';
 import { MatrixAdminEventUpdateRoomStateForAdminRoomsInput } from './dto/matrix.admin.rooms.dto.event.update.room.state.for.admin.rooms';
 import { MatrixAgentFactoryService } from '@src/domain/agent/agent-factory/matrix.agent.factory.service';
 import pkg  from '@nestjs/common';
+import { IMatrixUser } from '@src/domain/user/matrix.user.interface';
 const { Inject, Injectable } = pkg;
 
 @Injectable()
@@ -58,7 +58,7 @@ export class MatrixAdminRoomsService {
     );
   }
 
-  private async createMatrixClientForAdmin(adminUser: IOperationalMatrixUser) {
+  private async createMatrixClientForAdmin(adminUser: IMatrixUser) {
     const adminAgent =
       await this.agentFactoryService.createMatrixAgent(adminUser);
 
@@ -73,7 +73,7 @@ export class MatrixAdminRoomsService {
   private async getGlobalAdminUser(
     username: string,
     password: string
-  ): Promise<IOperationalMatrixUser> {
+  ): Promise<IMatrixUser> {
     const adminCommunicationsID =
       this.matrixUserAdapter.convertEmailToMatrixID(username);
     const adminExists = await this.adminUserService.isRegistered(
