+++ service/src/domain/room/matrix.room.options.ts
