+++ service/src/domain/user/matrix.user.interface.ts
@@ -2,8 +2,5 @@ export interface IMatrixUser {
   name: string;
   username: string;
   password: string;
-}
-
-export interface IOperationalMatrixUser extends IMatrixUser {
   accessToken: string;
 }
